+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+### 0.0.5
+  * Optimization of `EnvyProperty.updateValues`
+  * Refined `CanvasNode` support for mouse enter, leave, over, out and move
+  * Added `Sum`, `Diff`, `Product` and various other operation type NumberSources
+  * Added support for dashed lines
+  
 ### 0.0.4
   * `EnvyProperty.updateValues` no longer includes unavailable data when finishing
   * Bumped `polymer` dependency to `^1.0.0-rc.4`